+++ src/main/java/com/deepen/domain/WorkTmpDTO.java
@@ -1,5 +1,7 @@
 package com.deepen.domain;
 
+import java.math.BigDecimal;
+
 import lombok.Data;
 
 @Data
@@ -10,8 +12,8 @@ public class WorkTmpDTO {
 	private String work_end;
 	private String work_shift;
 	private String work_type;
-	private Integer work_time;
-	private Integer rest_time;
+	private BigDecimal  work_time;
+	private BigDecimal  rest_time;
 	// ------------------------------
 	private String shift_name;
 	private String type_name;
